requirements.txt
phidata
groq
google-generativeai
python-dotenv
fastapi
uvicorn
duckduckgo-search
packaging
